added checks for the phone number

+++ backend/src/main/java/ca/mcgill/ecse321/Mar1HotelSystem/service/CustomerService.java
@@ -3,7 +3,9 @@ package ca.mcgill.ecse321.Mar1HotelSystem.service;
 import java.util.List;
 
 import ca.mcgill.ecse321.Mar1HotelSystem.dao.*;
+import ca.mcgill.ecse321.Mar1HotelSystem.exception.Mar1HotelSystemException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import ca.mcgill.ecse321.Mar1HotelSystem.model.Customer;
@@ -47,22 +49,22 @@ public class CustomerService {
     public Customer createCustomer(String firstName, String lastName, String email, long phoneNumber, String password) {
         // Check if firstName is empty
         if (firstName == null || firstName.trim().isEmpty()) {
-            throw new IllegalArgumentException("The first name cannot be empty!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The first name cannot be empty!");
         }
         // Check if lastName is empty
         if (lastName == null || lastName.trim().isEmpty()) {
-            throw new IllegalArgumentException("The last name cannot be empty!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The last name cannot be empty!");
         }
         // Check if email is empty
         if (email == null || email.trim().isEmpty()) {
-            throw new IllegalArgumentException("The email cannot be empty!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The email cannot be empty!");
         }
         // Check if email is valid
         String emailTrimmed = email.trim();
         Pattern pattern = Pattern.compile("^(\\S+)@(\\S+)\\.((com)|(ca))$");
         Matcher matcher = pattern.matcher(emailTrimmed);
         if (!matcher.find()) {
-            throw new IllegalArgumentException("The email is invalid!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The email is invalid!");
         }
         // Check if email is already used
         // Note: why must customer be fore employee?
@@ -71,11 +73,15 @@ public class CustomerService {
                 || managerRepository.findManagerByEmail(emailTrimmed) != null
                 || generalUserRepository.findGeneralUserByEmail(emailTrimmed) != null
                 || accountRepository.findAccountByEmail(emailTrimmed) != null) {
-            throw new IllegalArgumentException("User using that email already exists!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "User using that email already exists!");
+        }
+        // Check if the phone number is above 0
+        if (phoneNumber <= 0) {
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The phone number must be above 0!");
         }
         // Check if password is empty
         if (password == null || password.trim().isEmpty()) {
-            throw new IllegalArgumentException("The password cannot be empty!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The password cannot be empty!");
         }
         // Create, save, and return the customer
         Customer customer = new Customer(firstName, lastName, email, phoneNumber, password);
@@ -86,31 +92,35 @@ public class CustomerService {
     public Customer updateCustomer(String newFirstName, String newLastName, long newPhoneNumber, String newPassword, String email) {
         // Check if firstName is empty
         if (newFirstName == null || newFirstName.trim().isEmpty()) {
-            throw new IllegalArgumentException("The first name cannot be empty!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The first name cannot be empty!");
         }
         // Check if lastName is empty
         if (newLastName == null || newLastName.trim().isEmpty()) {
-            throw new IllegalArgumentException("The last name cannot be empty!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The last name cannot be empty!");
         }
         // Check if email is empty
         if (email == null || email.trim().isEmpty()) {
-            throw new IllegalArgumentException("The email cannot be empty!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The email cannot be empty!");
         }
         // Check if email is valid
         String emailTrimmed = email.trim();
         Pattern pattern = Pattern.compile("^(\\S+)@(\\S+)\\.((com)|(ca))$");
         Matcher matcher = pattern.matcher(emailTrimmed);
         if (!matcher.find()) {
-            throw new IllegalArgumentException("The email is invalid!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The email is invalid!");
+        }
+        // Check if the phone number is above 0
+        if (newPhoneNumber <= 0) {
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The phone number must be above 0!");
         }
         // Check if password is empty
         if (newPassword == null || newPassword.trim().isEmpty()) {
-            throw new IllegalArgumentException("The password cannot be empty!");
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "The password cannot be empty!");
         }
         // Get Customer
         Customer customer = getCustomer(emailTrimmed);
         if (customer == null) {
-            throw new IllegalArgumentException("The customer does not exist!");
+            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND, "The customer does not exist!");
         }
         customer.setFirstName(newFirstName.trim());
         customer.setLastName(newLastName.trim());
@@ -122,11 +132,10 @@ public class CustomerService {
     @Transactional
     public boolean deleteCustomer(String email) {
         Customer customer = customerRepository.findCustomerByEmail(email);
-        try {
-            customerRepository.delete(customer);
-        } catch (Exception e) {
-            return false;
+        if (customer == null) {
+            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND, "The customer does not exist!");
         }
+        customerRepository.delete(customer);
         return true;
     }
